requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9.3 (tags/v3.9.3:e723086, Apr  2 2021, 11:01:03) [MSC v.1928 32 bit (Intel)]
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via django
build==1.0.3
    # via pip-tools
click==8.1.7
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
django==3.2.20
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-cors-headers==4.2.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.3.0
    # via -r requirements.in
drf-yasg==1.21.7
    # via -r requirements.in
importlib-metadata==6.8.0
    # via build
inflection==0.5.1
    # via drf-yasg
packaging==23.1
    # via
    #   build
    #   drf-yasg
pip-tools==7.3.0
    # via -r requirements.in
pyjwt==2.8.0
    # via djangorestframework-simplejwt
pymysql==1.1.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
python-decouple==3.8
    # via -r requirements.in
pytz==2023.3.post1
    # via
    #   django
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0.1
    # via drf-yasg
sqlparse==0.4.4
    # via django
termcolor==2.3.0
    # via -r requirements.in
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
typing-extensions==4.8.0
    # via asgiref
uritemplate==4.1.1
    # via drf-yasg
wheel==0.41.2
    # via pip-tools
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
